feat: added App.js to the separate folder
