+++ src/index.js
@@ -1,8 +1,8 @@
+import './index.scss';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
-import './index.scss';
-import App from './components/App';
+import App from './components/App/App';
 
 ReactDOM.render(
   <React.StrictMode>
